[SCD-64] rename controller folder and function. Change API message. Fix code style.
